requirements.txt
# --- Core UI & Visualization ---
streamlit>=1.40.0
streamlit-option-menu>=0.3.0
streamlit-extras>=0.4.0
plotly>=5.0.0
pandas>=2.0.0

# --- LangChain Ecosystem (v0.3+) ---
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-openai>=0.2.0
langchain-ollama>=0.1.0
langchain-chroma>=0.1.0
langchain-huggingface>=0.1.0

# --- Vector Store & Database ---
chromadb>=0.5.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# --- Document Extraction (OCR & Parsing) ---
kreuzberg>=0.4.0
pypdf>=3.0.0
python-docx>=1.0.0
beautifulsoup4>=4.12.0

# --- Model Connectivity ---
ollama>=0.3.0
openai>=1.0.0

# --- Data Validation ---
pydantic>=2.0.0
pydantic-settings>=2.0.0

# --- Utilities ---
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0

# --- Development & Testing ---
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
ruff>=0.1.0
mypy>=1.0.0

# --- Security ---
bandit>=1.7.0
safety>=2.3.0
